Reorganizing and Optimizing Front-End
So far I am catching up to Avry's Test-Bench version of the Front-End. So far I have implemented the original home page (font and background) and the Footer from the original HTML site. Similarly I have also optimized and corrected sections of the App.js/css file to better accommodate the changes that I have made. Now it is a little easier to implement other layout changes such as the planned Header, "Order Now!" section, "Reserve Seat" section, and lastly the Hero (the thing with the video in the background).