finished the edit page
in the reservation list page, you should now be able to press the edit button for a reservation and be taken to the edit reservation page for that user.
the version that I made is currently based off of Avry's current Reservation page using the same format and layout as hers.
changing the information for the reservation and pressing submit should send you back to the reservation list and the edits should not be shown on the list itself.
I also added the time of the reservation to the information shown in the reservation list page.